components/hydrosphere/requirements.txt
hydrosdk~=2.4.0rc
pandas==1.0.*
docker-image-py==0.1.*